new booking page, service requests attachments, project restructure and unit, integration, concurrent testing

+++ user-service/src/main/java/com/hse/userservice/internal/dto/CoworkingUserReadModelDto.java
@@ -6,7 +6,7 @@ public record CoworkingUserReadModelDto(
         Long userId,
         String name,
         LocalDate registeredAt,
-        Integer balance,
+        Long balance,
         Integer totalBookings,
         Integer unfinishedBookings
 ) {
